+++ packages/backend/src/api/controllers/status/discovery/props/getDashboardContracts.ts
@@ -1,24 +1,9 @@
-import {
-  ContractParameters,
-  ContractValue,
-  EthereumAddress,
-  ProjectParameters,
-  UpgradeabilityParameters,
-} from '@l2beat/shared'
-import { ethers } from 'ethers'
-import { isArray } from 'lodash'
+import { EthereumAddress, ProjectParameters } from '@l2beat/shared'
 
 import { DiscoveryConfig } from '../../../../../core/discovery/DiscoveryConfig'
-
-export interface DashboardContractFieldValue {
-  value: ContractValue
-  discoveryChild?: string
-}
-
-export interface DashboardContractField {
-  name: string
-  values?: DashboardContractFieldValue[]
-}
+import { getFunctions } from './utils/getFunctions'
+import { getUpgradeability } from './utils/getUpgradeability'
+import { DashboardContractField, getValues } from './utils/getValues'
 
 export interface DashboardContract {
   name: string
@@ -176,192 +161,6 @@ export function getDashboardContracts(
       proxyType: contract.upgradeability.type,
       isInitial: config.initialAddresses.includes(contract.address),
       discoveredBy,
-    }
-  })
-
-  return result
-}
-
-function getValues(
-  discovery: ProjectParameters,
-  contract: ContractParameters,
-  field: string,
-): {
-  value: ContractValue
-  discoveryChild?: string
-}[] {
-  const value = discovery.contracts.find((c) => c.address === contract.address)
-    ?.values?.[field]
-
-  if (value === undefined) {
-    return []
-  }
-
-  if (!isArray(value)) {
-    return [
-      {
-        value: value,
-        discoveryChild: getDiscoveryChild(discovery, contract, value),
-      },
-    ]
-  }
-
-  return value.map((v) => ({
-    value: v,
-    discoveryChild: getDiscoveryChild(discovery, contract, v),
-  }))
-}
-
-function getFunctions(
-  discovery: ProjectParameters,
-  address: EthereumAddress,
-): string[] {
-  const contract = discovery.contracts.find((c) => c.address === address)
-
-  if (contract === undefined) {
-    throw new Error(`Contract ${address.toString()} not found`)
-  }
-
-  if (contract.values === undefined) {
-    return []
-  }
-
-  const addresses = getAddresses(contract)
-
-  const functionNames: Set<string> = new Set<string>()
-
-  for (const address of addresses) {
-    const abi = Object.entries(discovery.abis).find(
-      ([a]) => a === address.toString(),
-    )
-    if (abi === undefined) {
-      throw new Error(`ABI for ${address.toString()} not found`)
-    }
-    abi[1].forEach((a) => {
-      if (a.includes('view')) {
-        functionNames.add(a)
-      }
-    })
-  }
-  const iface = new ethers.utils.Interface(Array.from(functionNames))
-
-  const functions = Object.entries(iface.functions)
-    .filter(([, fn]) => fn.stateMutability === 'view')
-    .map(([name]) => name)
-
-  return functions.sort()
-}
-
-function getAddresses(contract: ContractParameters): EthereumAddress[] {
-  const addresses: EthereumAddress[] = [contract.address]
-  if (contract.upgradeability.type === 'immutable') {
-    return addresses
-  }
-
-  if (contract.upgradeability.type === 'gnosis safe') {
-    addresses.push(contract.upgradeability.masterCopy)
-    return addresses
-  }
-
-  if (
-    contract.upgradeability.type === 'EIP1967 proxy' ||
-    contract.upgradeability.type === 'ZeppelinOS proxy' ||
-    contract.upgradeability.type === 'EIP897 proxy' ||
-    contract.upgradeability.type === 'call implementation proxy'
-  ) {
-    addresses.push(contract.upgradeability.implementation)
-    return addresses
-  }
-
-  if (contract.upgradeability.type === 'StarkWare proxy') {
-    addresses.push(contract.upgradeability.implementation)
-    if (contract.upgradeability.callImplementation) {
-      addresses.push(contract.upgradeability.callImplementation)
-    }
-    return addresses
-  }
-
-  if (contract.upgradeability.type === 'StarkWare diamond') {
-    addresses.push(contract.upgradeability.implementation)
-    addresses.push(...Object.values(contract.upgradeability.facets))
-    return addresses
-  }
-
-  if (contract.upgradeability.type === 'Arbitrum proxy') {
-    addresses.push(contract.upgradeability.adminImplementation)
-    addresses.push(contract.upgradeability.userImplementation)
-    return addresses
-  }
-
-  if (contract.upgradeability.type === 'new Arbitrum proxy') {
-    addresses.push(contract.upgradeability.implementation)
-    addresses.push(contract.upgradeability.adminImplementation)
-    addresses.push(contract.upgradeability.userImplementation)
-    return addresses
-  }
-
-  if (contract.upgradeability.type === 'resolved delegate proxy') {
-    addresses.push(contract.upgradeability.implementation)
-    return addresses
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-  if (contract.upgradeability.type === 'EIP2535 diamond proxy') {
-    addresses.push(...Object.values(contract.upgradeability.facets))
-    return addresses
-  }
-
-  throw new Error(
-    'Unhandled upgradeability type: ' + JSON.stringify(contract.upgradeability),
-  )
-}
-function getDiscoveryChild(
-  discovery: ProjectParameters,
-  contract: ContractParameters,
-  v: ContractValue,
-): string | undefined {
-  const i = discovery.contracts.find((c) => c.address.toString() === v)
-
-  if (i === undefined || i.address.toString() === contract.address.toString()) {
-    return undefined
-  }
-
-  return i.address.toString()
-}
-function getUpgradeability(
-  upgradeability: UpgradeabilityParameters,
-  discovery: ProjectParameters,
-  contract: ContractParameters,
-): DashboardContractField[] {
-  const result: DashboardContractField[] = []
-
-  Object.entries(upgradeability).forEach(([key, value]) => {
-    if (key === 'type' || value === undefined) {
-      return
-    }
-
-    if (!isArray(value)) {
-      result.push({
-        name: key,
-        values: [
-          {
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
-            value: value.toString(),
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-            discoveryChild: getDiscoveryChild(discovery, contract, value),
-          },
-        ],
-      })
-    } else {
-      result.push({
-        name: key,
-        values: value.map((v) => ({
-          // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-          value: v,
-          // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-          discoveryChild: getDiscoveryChild(discovery, contract, v),
-        })),
-      })
     }
   })
