+++ packages/backend/src/api/controllers/status/discovery/props/utils/getUpgradeability.ts
@@ -1,0 +1,49 @@
+import {
+  ContractParameters,
+  ProjectParameters,
+  UpgradeabilityParameters,
+} from '@l2beat/shared'
+import { isArray } from 'lodash'
+
+import { getDiscoveryChild } from './getDiscoveryChild'
+import { DashboardContractField } from './getValues'
+
+export function getUpgradeability(
+  upgradeability: UpgradeabilityParameters,
+  discovery: ProjectParameters,
+  contract: ContractParameters,
+): DashboardContractField[] {
+  const result: DashboardContractField[] = []
+
+  Object.entries(upgradeability).forEach(([key, value]) => {
+    if (key === 'type' || value === undefined) {
+      return
+    }
+
+    if (!isArray(value)) {
+      result.push({
+        name: key,
+        values: [
+          {
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
+            value: value.toString(),
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
+            discoveryChild: getDiscoveryChild(discovery, contract, value),
+          },
+        ],
+      })
+    } else {
+      result.push({
+        name: key,
+        values: value.map((v) => ({
+          // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+          value: v,
+          // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
+          discoveryChild: getDiscoveryChild(discovery, contract, v),
+        })),
+      })
+    }
+  })
+
+  return result
+}
