+++ packages/backend/src/api/controllers/status/discovery/props/utils/getValues.ts
@@ -1,0 +1,45 @@
+import {
+  ContractParameters,
+  ContractValue,
+  ProjectParameters,
+} from '@l2beat/shared'
+import { isArray } from 'lodash'
+
+import { getDiscoveryChild } from './getDiscoveryChild'
+
+export interface DashboardContractField {
+  name: string
+  values?: DashboardContractFieldValue[]
+}
+
+export interface DashboardContractFieldValue {
+  value: ContractValue
+  discoveryChild?: string
+}
+
+export function getValues(
+  discovery: ProjectParameters,
+  contract: ContractParameters,
+  field: string,
+): DashboardContractFieldValue[] {
+  const value = discovery.contracts.find((c) => c.address === contract.address)
+    ?.values?.[field]
+
+  if (value === undefined) {
+    return []
+  }
+
+  if (!isArray(value)) {
+    return [
+      {
+        value: value,
+        discoveryChild: getDiscoveryChild(discovery, contract, value),
+      },
+    ]
+  }
+
+  return value.map((v) => ({
+    value: v,
+    discoveryChild: getDiscoveryChild(discovery, contract, v),
+  }))
+}
